Move view helper functions to handlebars
- Create /views/helpers/index.js.
- Move firstLetter and color contrast function to helper.
- Remove functions from crud.js.
- Update list.hbs and view.hbs to use Handlebar helpers.
