+++ views/pwas/list.hbs
@@ -27,7 +27,7 @@
                     <div class="media-left">
                         <svg width="64" height="64">
                             <rect width="64" height="64" stroke-width="4" fill="{{backgroundColor}}" />
-                            <text x="32" y="32" fill="{{fgcolor}}" alignment-baseline="middle" text-anchor="middle" font-size="32px">{{firstletter}}</text>                            
+                            <text x="32" y="32" fill="{{contrastColor backgroundColor}}" alignment-baseline="middle" text-anchor="middle" font-size="32px">{{firstLetter name}}</text>
                         </svg>                
                     </div>
                     <div class="media-body">
