+++ views/pwas/view.hbs
@@ -33,7 +33,7 @@
                 <div class="media-left">
                     <svg width="256" height="256">
                         <rect width="100%" height="100%" stroke-width="4" fill="{{pwa.backgroundColor}}" />
-                        <text x="50%" y="50%" fill="{{pwa.fgcolor}}" alignment-baseline="middle" text-anchor="middle" font-size="128px">{{pwa.firstletter}}</text>                            
+                        <text x="50%" y="50%" fill="{{contrastColor pwa.backgroundColor}}" alignment-baseline="middle" text-anchor="middle" font-size="128px">{{firstLetter pwa.name}}</text>
                     </svg>                    
                 </div>
                 <div class="media-body">
